+++ src/BuiltInPackages/LinqSTG/LuaSTGEditorSharpV2.Package.LinqSTG.Window/ViewModel/LinqSTGNodeInputViewModel.cs
@@ -313,9 +313,9 @@ namespace LuaSTGEditorSharpV2.Package.LinqSTG.Windows.ViewModel
             };
         }
 
-        public static LinqSTGNodeInputViewModel<Contextual<IParametric<int, Vector2>>?> Movement(string? name = null)
+        public static LinqSTGNodeInputViewModel<Contextual<IParametric<float, Vector2>>?> Movement(string? name = null)
         {
-            return new LinqSTGNodeInputViewModel<Contextual<IParametric<int, Vector2>>?>
+            return new LinqSTGNodeInputViewModel<Contextual<IParametric<float, Vector2>>?>
             {
                 Name = name,
                 Port = new LinqSTGPortViewModel(PortColor.Movement),
